requirements.txt
langchain>=0.2.27
langchain-groq>=0.1.0
langchain-astradb>=0.2.0
langchain-core>=0.2.27
langchain-community>=0.2.0
langgraph>=0.2.0
python-dotenv
pydantic-settings
flask
flask-cors
requests
huggingface-hub
astrapy>=1.1.0,<2.0
numpy
pandas
PyMuPDF
pymongo>=4.0
